epileptech-api/requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
pymongo==4.6.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
bcrypt==4.1.2
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.2
tensorflow==2.15.0
transformers==4.38.2
torch==2.2.1
torch-geometric==2.4.0
python-socketio==5.11.1
eventlet==0.35.2
matplotlib==3.8.0
fpdf2==2.8.3
mne==1.5.1
openai==1.76.0
